requirements.txt
# Self-Driving Car Project Dependencies
# Industry-standard packages with pinned versions

# Core ML/AI Framework
torch>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# GUI Framework
kivy>=2.1.0,<3.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0

# Configuration Management
PyYAML>=6.0,<7.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.0,<4.0.0

# Code Quality
black>=23.7.0,<24.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.5.0,<2.0.0

# Type Stubs
types-PyYAML>=6.0.12

# Documentation
sphinx>=7.1.0,<8.0.0
sphinx-rtd-theme>=1.3.0

# Optional: TensorBoard for training visualization
tensorboard>=2.13.0,<3.0.0
